wip(sync): POC of using Redis for pubsub
