--- umap/sync/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class UmapConfig(AppConfig):
-    name = "umap.sync"
-    verbose_name = "uMap Sync"
